Try all primitives eagerly
More importantly, use primitive indices as identifiers for eager sends instead of selectors

+++ trufflesqueak/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/model/CompiledCodeObject.java
@@ -309,7 +309,7 @@ public abstract class CompiledCodeObject extends SqueakObject {
         return hasPrimitive;
     }
 
-    public long primitiveIndex() {
+    public int primitiveIndex() {
         if (hasPrimitive && bytes.length >= 3) {
             return Byte.toUnsignedInt(bytes[1]) + (Byte.toUnsignedInt(bytes[2]) << 8);
         } else {
